+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## Master
 
+## 0.8.2
+
 * Support multiple Danger instances with `--dangerId` - marcelofabri
 * Add base request source so services other than GitHub could be used with Danger. - justMaku
-* Don't validate CI sources that don't expose all required environment variables.
-* Add support for TeamCity CI
+* Don't validate CI sources that don't expose all required environment variables.  - justMaku
+* Add support for TeamCity CI - rbuussyghin
 
 ## 0.8.1